changing view port and canvas dimensions to prevent integer devision

+++ Constants.h
@@ -5,11 +5,11 @@
 
 #include "Vec3.h"
 
-float CANVAS_WIDTH = 500;
-float CANVAS_HEIGHT = 500;
+int CANVAS_WIDTH = 500;
+int CANVAS_HEIGHT = 500;
 
-float VIEWPORT_HEIGHT = 1;
-float VIEWPORT_WIDTH = 1;
+int VIEWPORT_HEIGHT = 1;
+int VIEWPORT_WIDTH = 1;
 
 int PROJECTTION_PLANE_D = 1;
 Vec3 origin = Vec3(0, 0, 0);
